Modified two redux slices of doctors and two table components also Updated two administrator dashboard pages
